move clusterUtilization into SingularityClusterUtilization
