+++ SingularityService/src/main/java/com/hubspot/singularity/config/SingularityConfiguration.java
@@ -79,6 +79,8 @@ public class SingularityConfiguration extends Configuration {
 
   private int numUsageToKeep = 15;
 
+  private double minUnderUtilizedPct = 0.05;
+
   private long cleanupEverySeconds = 5;
 
   private long checkQueuedMailsEveryMillis = TimeUnit.SECONDS.toMillis(15);
@@ -1431,6 +1433,15 @@ public class SingularityConfiguration extends Configuration {
 
   public void setNumUsageToKeep(int numUsageToKeep) {
     this.numUsageToKeep = numUsageToKeep;
+  }
+
+  public double getMinUnderUtilizedPct() {
+    return minUnderUtilizedPct;
+  }
+
+  public SingularityConfiguration setMinUnderUtilizedPct(double minUnderUtilizedPct) {
+    this.minUnderUtilizedPct = minUnderUtilizedPct;
+    return this;
   }
 
   public long getCacheOffersForMillis() {
