+++ SingularityService/src/main/java/com/hubspot/singularity/resources/UsageResource.java
@@ -1,24 +1,17 @@
 package com.hubspot.singularity.resources;
 
 import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.Iterator;
 import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
 
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
-import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.MediaType;
 
 import com.google.common.base.Optional;
 import com.google.inject.Inject;
-import com.hubspot.singularity.ClusterUtilization;
-import com.hubspot.singularity.RequestUtilization;
-import com.hubspot.singularity.SingularityDeploy;
+import com.hubspot.singularity.SingularityClusterUtilization;
 import com.hubspot.singularity.SingularityService;
 import com.hubspot.singularity.SingularitySlave;
 import com.hubspot.singularity.SingularitySlaveUsage;
@@ -28,12 +21,10 @@ import com.hubspot.singularity.SingularityTaskCurrentUsageWithId;
 import com.hubspot.singularity.SingularityTaskId;
 import com.hubspot.singularity.SingularityTaskUsage;
 import com.hubspot.singularity.WebExceptions;
-import com.hubspot.singularity.data.DeployManager;
 import com.hubspot.singularity.data.SlaveManager;
 import com.hubspot.singularity.data.TaskManager;
 import com.hubspot.singularity.data.UsageManager;
 import com.wordnik.swagger.annotations.Api;
-import com.wordnik.swagger.annotations.ApiParam;
 
 @Path(UsageResource.PATH)
 @Produces({ MediaType.APPLICATION_JSON })
@@ -45,14 +36,12 @@ public class UsageResource {
   private final UsageManager usageManager;
   private final TaskManager taskManager;
   private final SlaveManager slaveManager;
-  private final DeployManager deployManager;
 
   @Inject
-  public UsageResource(UsageManager usageManager, TaskManager taskManager, SlaveManager slaveManager, DeployManager deployManager) {
+  public UsageResource(UsageManager usageManager, TaskManager taskManager, SlaveManager slaveManager) {
     this.usageManager = usageManager;
     this.taskManager = taskManager;
     this.slaveManager = slaveManager;
-    this.deployManager = deployManager;
   }
 
   @GET
@@ -92,90 +81,9 @@ public class UsageResource {
 
   @GET
   @Path("/cluster/utilization")
-  public ClusterUtilization getClusterUtilization(@ApiParam("For under-utilized requests, only consider those that are under-utilized greater than or equal to this percentage") @QueryParam("minUnderUtilizedPct") double minUnderUtilizedPct) {
-    List<SingularityTaskId> tasks = usageManager.getTasksWithUsage().stream().map(SingularityTaskId::valueOf).collect(Collectors.toList());
+  public SingularityClusterUtilization getClusterUtilization() {
+    WebExceptions.checkNotFound(usageManager.getClusterUtilization().isPresent(), "No cluster utilization has been saved yet");
 
-    Map<String, RequestUtilization> utilizationPerRequestId = new HashMap<>();
-
-    for (SingularityTaskId task : tasks) {
-      RequestUtilization requestUtilization = utilizationPerRequestId.getOrDefault(task.getRequestId(), new RequestUtilization(task.getRequestId(), task.getDeployId()));
-
-      usageManager.getTaskUsage(task.getId()).forEach(usage -> {
-        requestUtilization.addCpu(usage.getCpuSeconds());
-        requestUtilization.addMemBytes(usage.getMemoryRssBytes());
-        requestUtilization.incrementTaskCount();
-      });
-
-      utilizationPerRequestId.put(task.getRequestId(), requestUtilization);
-    }
-
-    return getClusterUtilization(utilizationPerRequestId, minUnderUtilizedPct);
-  }
-
-  private ClusterUtilization getClusterUtilization(Map<String, RequestUtilization> utilizationPerRequestId, double minUnderUtilizedPct) {
-    int numRequestsWithUnderUtilizedCpu = 0;
-    int numRequestsWithOverUtilizedCpu = 0;
-    int numRequestsWithUnderUtilizedMemBytes = 0;
-
-    double totalUnderUtilizedCpu = 0;
-    double totalOverUtilizedCpu = 0;
-    long totalUnderUtilizedMemBytes = 0;
-
-    double maxUnderUtilizedCpu = Double.MIN_VALUE;
-    double maxOverUtilizedCpu = Double.MIN_VALUE;
-    long maxUnderUtilizedMemBytes = Long.MIN_VALUE;
-
-    double minUnderUtilizedCpu = Double.MAX_VALUE;
-    double minOverUtilizedCpu = Double.MAX_VALUE;
-    long minUnderUtilizedMemBytes = Long.MAX_VALUE;
-
-
-    for (Iterator<Map.Entry<String, RequestUtilization>> it = utilizationPerRequestId.entrySet().iterator(); it.hasNext(); ) {
-      RequestUtilization utilization = it.next().getValue();
-      Optional<SingularityDeploy> maybeDeploy = deployManager.getDeploy(utilization.getRequestId(), utilization.getDeployId());
-
-      if (maybeDeploy.isPresent() && maybeDeploy.get().getResources().isPresent()) {
-        boolean includeUtilization = true;
-        long memoryBytesReserved = (long) (maybeDeploy.get().getResources().get().getMemoryMb() * SingularitySlaveUsage.BYTES_PER_MEGABYTE);
-        double cpuReserved = maybeDeploy.get().getResources().get().getCpus();
-
-        double unusedCpu = cpuReserved - utilization.getAvgCpuUsed();
-        long unusedMemBytes = memoryBytesReserved - utilization.getMemBytesTotal();
-
-        if (unusedCpu / cpuReserved >= minUnderUtilizedPct) {
-          numRequestsWithUnderUtilizedCpu++;
-          totalUnderUtilizedCpu += unusedCpu;
-          maxUnderUtilizedCpu = Math.max(unusedCpu, maxUnderUtilizedCpu);
-          minUnderUtilizedCpu = Math.min(unusedCpu, minUnderUtilizedCpu);
-        } else if (unusedCpu < 0) {
-          double overusedCpu = Math.abs(unusedCpu);
-
-          numRequestsWithOverUtilizedCpu++;
-          totalOverUtilizedCpu += overusedCpu;
-          maxOverUtilizedCpu = Math.max(overusedCpu, maxOverUtilizedCpu);
-          minOverUtilizedCpu = Math.min(overusedCpu, minOverUtilizedCpu);
-        } else {
-          includeUtilization = false;
-        }
-
-        if (unusedMemBytes / memoryBytesReserved >= minUnderUtilizedPct) {
-          numRequestsWithUnderUtilizedMemBytes++;
-          totalUnderUtilizedMemBytes += unusedMemBytes;
-          maxUnderUtilizedMemBytes = Math.max(unusedMemBytes, maxUnderUtilizedMemBytes);
-          minUnderUtilizedMemBytes = Math.min(unusedMemBytes, minUnderUtilizedMemBytes);
-        } else if (!includeUtilization) {
-          it.remove();
-        }
-      }
-    }
-
-    double avgUnderUtilizedCpu = totalUnderUtilizedCpu / numRequestsWithUnderUtilizedCpu;
-    double avgOverUtilizedCpu = totalOverUtilizedCpu / numRequestsWithOverUtilizedCpu;
-    double avgUnderUtilizedMemBytes = totalUnderUtilizedMemBytes / numRequestsWithUnderUtilizedMemBytes;
-
-    return new ClusterUtilization(new ArrayList<>(utilizationPerRequestId.values()), numRequestsWithUnderUtilizedCpu, numRequestsWithOverUtilizedCpu,
-        numRequestsWithUnderUtilizedMemBytes, totalUnderUtilizedCpu, totalOverUtilizedCpu, totalUnderUtilizedMemBytes, avgUnderUtilizedCpu,
-        avgOverUtilizedCpu, avgUnderUtilizedMemBytes, maxUnderUtilizedCpu, maxOverUtilizedCpu, maxUnderUtilizedMemBytes, minUnderUtilizedCpu,
-        minOverUtilizedCpu, minUnderUtilizedMemBytes);
+    return usageManager.getClusterUtilization().get();
   }
 }
